move integration test environment to testutils
